xssgg.py
import requests
from playwright.sync_api import sync_playwright
import typer

app = typer.Typer()


@app.command()
def exploit():
    url = 'https://google-gruyere.appspot.com/527569850200309825989661872787286261006/'

    with sync_playwright() as playwright:
        browser = playwright.chromium.launch()
        page = browser.new_page()
        page.goto(url)

        # Perform the automation exploit using Playwright

        browser.close()

    typer.echo("Exploit completed successfully!")


@app.command()
def other_command():
    # Implement other commands using Typer
    pass


if __name__ == "__main__":
    app()