test_xssgg.py
import pytest
from playwright.sync_api import sync_playwright

# Test case for the exploit functionality
def test_exploit():
    url = 'https://google-gruyere.appspot.com/527569850200309825989661872787286261006/'

    with sync_playwright() as playwright:
        browser = playwright.chromium.launch()
        page = browser.new_page()
        page.goto(url)

        # Perform the automation exploit using Playwright

        browser.close()

    # Add appropriate assertions to verify the exploit functionality
    assert True

# Test case for other command functionality
def test_other_command():
    # Implement test case for other command functionality using Typer or other techniques
    assert True